requirements.txt
pandas
streamlit
altair
python-dotenv
openai

